Conflicts:
	Assets/Models/LeafTree50Shape.dae
	Assets/Models/PaintedIslandBigShape.dae
	Assets/Models/PaintedIslandShape.dae
	Assets/Models/PaintedIslandShape1.dae
	Assets/Models/PaintedIslandShape2.dae
	Assets/Models/PaintedIslandShape3.dae
	Assets/Models/RootBridgeShape.dae
	Assets/Models/RootPlatformShape.dae
	Assets/Models/RootPlatformShape1.dae
	Assets/Models/RootPlatformShape3.dae
	Assets/Models/RootPlatformShape33.dae
	Assets/Models/RootPlatformShape35.dae
	Assets/Models/RootPlatformShape36.dae
	Assets/Models/RootPlatformShape7.dae
	Assets/Models/RootPlatformShape8.dae
	Assets/Models/RootsShape.dae
	Assets/Models/RootsShape1.dae
	Assets/Models/RootsShape2.dae
	Assets/Models/RootsShape3.dae
	Assets/Models/RootsShape4.dae
	Assets/Models/RuneGlowStone2Shape.dae
	Assets/Models/RuneGlowStone7Shape.dae
	Assets/Models/RuneGlowStoneShape.dae

+++ RootSystems/Include/Shapes/Rectangle.h
@@ -6,10 +6,10 @@ namespace RootForce
 	struct Rectangle
 	{
 		Rectangle();
-		Rectangle(int p_x, int p_y, int p_width, int p_height);
+		Rectangle(float p_x, float p_y, float p_width, float p_height);
 	
 		bool IsPointContained(glm::vec2 p_point);
 
-		int m_x, m_y, m_width, m_height;
+		float m_x, m_y, m_width, m_height;
 	};
 }
